refactor(query/tests): splash sql parser tests
