+++ query/tests/it/sql/parsers/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,8 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod optimizer;
-mod parsers;
-mod plan_parser;
-mod sql_parser;
-mod statements;
+mod parser_admin;
+mod parser_copy;
+mod parser_database;
+mod parser_optimize;
+mod parser_show;
+mod parser_stage;
+mod parser_table;
+mod parser_udf;
+mod parser_use;
+mod parser_user;
